requirements.txt
opencv-python
mediapipe
pyserial
requests
numpy